refactor(mespapiers): Extract ConnectorItem into HarvestBanner file
