+++ packages/cozy-mespapiers-lib/src/components/Papers/PapersList.jsx
@@ -2,11 +2,9 @@ import React, { useMemo, useState } from 'react'
 import PropTypes from 'prop-types'
 import flag from 'cozy-flags'
 
-import { LaunchTriggerCard } from 'cozy-harvest-lib'
-import { useClient, useQuery } from 'cozy-client'
+import { useClient } from 'cozy-client'
 import Button from 'cozy-ui/transpiled/react/Buttons'
 import { useI18n } from 'cozy-ui/transpiled/react/I18n'
-import Divider from 'cozy-ui/transpiled/react/MuiCozyTheme/Divider'
 
 import PaperLine from '../Papers/PaperLine'
 import { useModal } from '../Hooks/useModal'
@@ -18,49 +16,7 @@ import { rename } from '../Actions/Items/rename'
 import { viewInDrive } from '../Actions/Items/viewInDrive'
 import { makeActionVariant, makeActions } from '../Actions/utils'
 import { useMultiSelection } from '../Hooks/useMultiSelection'
-import {
-  buildTriggersQueryByConnectorSlug,
-  buildConnectorsQueryById
-} from '../../helpers/queries'
-
-const ConnectorItem = ({ papers }) => {
-  const { createdByApp: connectorSlug, sourceAccount: connectorAccount } =
-    papers?.list?.[0]?.cozyMetadata || {}
-  const queryTriggers = buildTriggersQueryByConnectorSlug(
-    connectorSlug,
-    Boolean(connectorSlug)
-  )
-  const { data: triggers } = useQuery(
-    queryTriggers.definition,
-    queryTriggers.options
-  )
-  const trigger = triggers?.find(
-    trigger => trigger.message.account === connectorAccount
-  )
-
-  const queryKonnector = buildConnectorsQueryById(
-    `io.cozy.konnectors/${connectorSlug}`,
-    Boolean(trigger)
-  )
-
-  const { data: konnectors } = useQuery(
-    queryKonnector.definition,
-    queryKonnector.options
-  )
-  const konnector = konnectors?.[0]
-
-  if (!konnector) return null
-
-  return (
-    <>
-      <LaunchTriggerCard
-        flowProps={{ initialTrigger: trigger, konnector }}
-        konnectorRoot={`harvest/${connectorSlug}/accounts/${connectorAccount}`}
-      />
-      <Divider />
-    </>
-  )
-}
+import HarvestBanner from './HarvestBanner'
 
 const PapersList = ({ papers, isLast }) => {
   const client = useClient()
@@ -104,7 +60,7 @@ const PapersList = ({ papers, isLast }) => {
   return (
     <>
       {flag('harvest.inappconnectors.enabled') && (
-        <ConnectorItem papers={papers} />
+        <HarvestBanner papers={papers} />
       )}
       {papers.list.map(
         (paper, idx) =>
